apps/backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main", "seed_agent", "story_agent", "fal_client", "image_prompt_agent"]

[project]
name = "storyapp-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sse-starlette",
    "ai-definitions",
    "pydantic-ai",
    "openai",
    "python-dotenv",
    "httpx",
]
